+++ src/app/components/Loader/index.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react'
+
+import style from './Loader.css'
+
+const Loader = () => (
+  <div className={style.loader}>
+  </div>
+)
+
+export default Loader
